The bug is solved
Some books are too big, so the image-downloader starts like 500 threads
trying to download pics as soon as possible. As a result the server
rejects the connection and some pngs are not downloaded. Fortunately
pdf builder notices that and rejects the book.

So, I tried to find library that do not have threaded download, but I
could not do it. They had their own limitations.

I solved it by simply checking if the png was already downloaded. If it
was then it does not download. So we simply run the script twice and it
downloads pages which were rejected first time.

Cherrypicked several other things.